+++ common/src/main/java/net/picopress/mc/mods/zombietactics2/ai/goals/move/SelectiveFloatGoal.java
@@ -1,5 +1,6 @@
-package net.picopress.mc.mods.zombietactics2.goals.move;
+package net.picopress.mc.mods.zombietactics2.ai.goals.move;
 
+import net.picopress.mc.mods.zombietactics2.Config;
 import net.picopress.mc.mods.zombietactics2.mixin.ZombieMisc;
 
 import net.minecraft.world.entity.LivingEntity;
@@ -41,17 +42,18 @@ public class SelectiveFloatGoal extends FloatGoal {
         // selectively float
         // and, zombies want to breathe, but not want to be drowned
         if(misc.getInWaterTime() > calculateBreath(target)) needBreathe = true;
-        return super.canUse() && (!target.isInWater() || mob.getBlockY() - target.getBlockY() < 1 || needBreathe);
+        return super.canUse() && (mob.getBlockY() - target.getBlockY() < 1 || needBreathe);
     }
 
     @Override
     public boolean canContinueToUse() {
         if(misc.getInWaterTime() < 15) needBreathe = false;
-        return super.canUse() || needBreathe;
+        return super.canUse() && needBreathe;
     }
 
     // calculate the cost of swimming
-    private int calculateBreath(@NotNull LivingEntity target) {
-        return 600 - (mob.getBlockY() - target.getBlockY()) * 30;
+    private double calculateBreath(@NotNull LivingEntity target) {
+        // thonk
+        return 600 - (mob.getBlockY() - target.getBlockY()) * 75 / Config.swimSpeed;
     }
 }